Update fast_bilateral_solver.py
Optimized the part of the makeBilateralGrid() function that computes the component blur matrices to run faster.

Completed solving for A and b and backprojecting the solution onto an image plane. The output image also has color.

Added a line where the cv2 bilateral filter is used so we can compare ours to it.